♻️ Move room login management to room router
