+++ src/utils/use-safe-effect.ts
@@ -1,0 +1,15 @@
+import React from 'react'
+
+export const useSafeEffect = (callback: () => void) => {
+  const isMounted = React.useRef(true)
+
+  React.useEffect(() => () => {
+    isMounted.current = false
+  })
+
+  React.useEffect(() => {
+    if (isMounted.current) {
+      return callback()
+    }
+  }, [callback])
+}
